+++ src/components/Loader/styles.js
@@ -1,19 +1,5 @@
 import { StyleSheet } from 'react-native';
 
-import styleConstants from '../../styleConstants';
-
-const styles = StyleSheet.create({
-  wrapper: {
-    position: 'absolute',
-    bottom: 0,
-    left: 0,
-    right: 0,
-    height: 5,
-  },
-  container: {
-    width: 100,
-    height: 5,
-  },
-});
+const styles = StyleSheet.create({});
 
 export default styles;